% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DART_Code.R
\name{A.tree.mult}
\alias{A.tree.mult}
\title{Stage 1: build aggregation tree}
\usage{
A.tree.mult(grids, Dist0, Mgroup = 3)
}
\arguments{
\item{grids}{(vector) vector of maximum distance on layers 2,...,L. The maximum distance on the first layer is always 0.}

\item{Dist0}{(matrix) Optional distance matrix. Default is NULL.}

\item{Mgroup}{(integer) maximum cardinality of the child node set of the nodes in the aggregation tree.}
}
\value{
A list including constructed tree and distance matrix: 1. Llist: (list of matrixs) Aggregation tree presented in the form of matrix list; 2, Dist0: (matrix) distance matrix
}
\description{
Stage 1: build aggregation tree
}
