% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{accessors}
\alias{accessors}
\alias{product}
\alias{simname}
\alias{files}
\alias{bands}
\alias{iters}
\alias{variables}
\alias{variablesRB3D}
\alias{typeNums}
\alias{imageType}
\alias{imageNo}
\title{Access object information}
\usage{
product(x)

simname(x)

files(x)

bands(x)

iters(x)

variables(x)

variablesRB3D(x)

typeNums(x)

imageType(x)

imageNo(x)
}
\description{
Generic functions to access information from the objects with classes
defined in this package
}
