% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorial.R
\name{all_lcs}
\alias{all_lcs}
\title{all_lcs}
\usage{
all_lcs(p, q)
}
\arguments{
\item{p}{integer vector.}

\item{q}{integer vector.}
}
\value{
(k x m) matrix of LCS's, where k = size of LCS and m = number of LCS.
}
\description{
Computes all LCS's between two integer vectors.
}
