% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/around_the_clock.R
\name{around_the_clock}
\alias{around_the_clock}
\title{around the clock training session}
\usage{
around_the_clock(section)
}
\arguments{
\item{section}{which section to score in: "s" for singles (default), "d" for doubles or "t" for trebles}
}
\description{
This function is used to play around the clock in training session
}
\author{
Matteo Miotto
}
