% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\docType{methods}
\name{as.data.frame,SimulationData-method}
\alias{as.data.frame,SimulationData-method}
\title{as.data.frame}
\usage{
\S4method{as.data.frame}{SimulationData}(x, as.tibble = TRUE)
}
\arguments{
\item{x}{SimulationData.}

\item{as.tibble}{Return as a tibble-type data frame?}
}
\value{
data.frame or tibble
}
\description{
as.data.frame
}
