% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{bandInfo}
\alias{bandInfo}
\title{bandInfo}
\usage{
bandInfo(x = "SimulationFiles")
}
\arguments{
\item{x}{sF \link{SimulationFiles-class}}
}
\description{
get full information on wavelengths for each band
}
