% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_persistence.R
\name{bigraded_betti}
\alias{bigraded_betti}
\title{Bigraded Betti numbers}
\usage{
bigraded_betti(
  x,
  f,
  H = 1L,
  xbin = 10L,
  ybin = 10L,
  max_dist = NULL,
  rivet_path = "~/rivet",
  save.path = NULL,
  xreverse = FALSE,
  yreverse = FALSE
)
}
\arguments{
\item{x}{point set.}

\item{f}{function values on \code{x}.}

\item{H}{homology dimension to compute.}

\item{xbin}{number of bins to coarsen \code{f} with.}

\item{ybin}{number of bins to coarsen the rips filtration values on \code{x} with.}

\item{max_dist}{maximum diameter to consider for rips filtration.}

\item{rivet_path}{directory where \emph{rivet_console} is located. Defaults to "~/rivet". Must be valid.}

\item{save.path}{if supplied, an input file giving the parameters to run RIVET is saved here.}

\item{xreverse}{unused.}

\item{yreverse}{unused.}
}
\description{
Returns the bigraded Betti numbers of a bifiltration.
}
\details{
The function uses the RIVET software to compute the bigraded betti numbers of a bifiltration 
built from point cloud \code{x} and function \code{f}.
}
