% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_persistence.R
\name{bigraded_betti2}
\alias{bigraded_betti2}
\title{Bigraded_betti2}
\usage{
bigraded_betti2(
  x,
  f1,
  f2,
  H = 1L,
  xbin = 10L,
  ybin = 10L,
  rivet_path = "~/rivet",
  save.path = NULL,
  xreverse = FALSE,
  yreverse = FALSE
)
}
\arguments{
\item{x}{list of simplices in filtration order}

\item{f1}{first filter function}

\item{f2}{second filter function}
}
\description{
Bigraded_betti2
}
