% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_persistence.R
\name{bipersistence}
\alias{bipersistence}
\title{bipersistence}
\usage{
bipersistence(
  x,
  f,
  H = 0L,
  xbin = 10,
  ybin = 10,
  max_dist = 1,
  rivet_path = "~/rivet",
  ...
)
}
\arguments{
\item{x}{(m x d) matrix of m points representing a point set.}

\item{f}{function on \code{x} to create bifiltration with.}
}
\description{
Computes the information needed for 2D persistence
}
