% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcDART.R
\name{CalcDART}
\alias{CalcDART}
\title{Calculates DART}
\usage{
CalcDART(DART_obj, x, y = NULL, nControl = 100)
}
\description{
# TODO: somewhere in the CI calculation loop it is silently failing on some
#       iterations. This skips the iteration, but very slowly, and re-running
#       CalcDART when % finished is less than numValues takes just as long
#       but does not add completed pixels. Would be good if this was sped up.

End-user interface for calculating DART

Replaces dart_fn and get_art

# notes from SF:

 Generate dart pixels

 X = x coordinate in EPSG 5070
  Y = y coordinate in EPSG 5070
   buffer = distance to buffer target in m
   searchRadius = distance to search for matches (m) == DART_obj@search_radius
   nControl = number of matches to look for
   targetDim = edge length in pixels of target area (1x1, 2x2, 3x3)
   x and y within bounding area
   buffer >= 0
   searchRadius >= 0 & > buffer
   nControl > 0
   targetDim > 0 & less than MAX
}
\details{
NA
}
