% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closing_3.R
\name{closing_3}
\alias{closing_3}
\title{closing in 3 darts}
\usage{
closing_3(level)
}
\arguments{
\item{level}{"beginner" (default), "intermediate or "advanced"}
}
\description{
This function is used to train in closures with only 3 darts
}
\author{
Matteo Miotto
}
