% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closing_6.R
\name{closing_6}
\alias{closing_6}
\title{closing exercise}
\usage{
closing_6(level)
}
\arguments{
\item{level}{"beginner" (default), "intermediate or "advanced"}
}
\description{
This function is used to train in closures with 3 or 6 darts
}
\author{
Matteo Miotto
}
