% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DART.R
\docType{class}
\name{DART}
\alias{DART}
\alias{DART-class}
\alias{initialize,DART-method}
\title{S4 definition for DART-class objects}
\usage{
\S4method{initialize}{DART}(.Object, ...)
}
\description{
NA
}
\details{
NA
}
\section{Slots}{

\describe{
\item{\code{ID}}{Identification/name of this DART object. See details.}

\item{\code{years}}{Years that response variable is desired. Optional.}

\item{\code{spatial_directory}}{Directory path containing spatial files. Optional.}

\item{\code{intermediate_directory}}{Directory to save intermediate files}

\item{\code{results_directory}}{Directory path for results files. Optional.}

\item{\code{save_intermediates}}{Logical flag - save intermediate rasters?}

\item{\code{resume}}{Logical flag - resume at last intermediate?}

\item{\code{diagnostic_plots}}{Logical flag - plot diagnostic plots for the GIS work?}

\item{\code{synthetic_control}}{Logical flag - do synthetic control?}

\item{\code{skip_CI_for_debug}}{Logical flag - skip CI for debug?}

\item{\code{skip_topo_for_debug}}{Logical flag - skip CI for debug?}

\item{\code{target_prop}}{Proportion of pixels in each polygon to select for CI
calculations. Must be > 0 and <= 1, where 1 is all pixels}

\item{\code{ref_prop}}{Proportion of pixels in reference area to subset. Useful for debugging.}

\item{\code{CI_prop}}{Proportion of eligible pixels to use for CI analysis}

\item{\code{seed}}{value for set.seed()}

\item{\code{test_old_ART}}{Logical flag for testing against old ART routine results}

\item{\code{search_radius}}{Radius in meters to search for reference pixels.}

\item{\code{target_radius}}{Radius in meters of target area}

\item{\code{buffer_radius}}{Radius in meters of target-reference buffer zone}

\item{\code{n_topographic_matches}}{Number of topographic matches to search for.
Optional. See details.}

\item{\code{n_cores}}{Number of cores to use for parallelization. Optional.}

\item{\code{NLCD_filter}}{Numeric of NLCD codes to filter by. Optional.}

\item{\code{mask}}{Study area mask. SpatVector. Required.}

\item{\code{nlcd_mask}}{Land cover mask. Filename. Optional.}

\item{\code{other_masks}}{Other mask layers. List of filenames. Optional.}

\item{\code{target_masks}}{Masks just for target layer. Optional.}

\item{\code{other_controls}}{Other control covariates. List of filenames. Optional.}

\item{\code{response_variables}}{List of response variable filenames}

\item{\code{result_data}}{Dataframe with result summary data}
}}

