% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart_centroid.R
\name{dart_centroid}
\alias{dart_centroid}
\title{sf::st_centroid() wrapper with optional CRS transformation}
\usage{
dart_centroid(sf, target_crs = 4326, output_crs = 4326)
}
\arguments{
\item{sf}{sf object}

\item{target_crs}{CRS which the centroid calculation will be performed
with (optional)}

\item{output_crs}{CRS which the function will return}
}
\value{
original sf object with new centroid column
}
\description{
sf::st_centroid() wrapper with optional CRS transformation
}
\examples{
if (rlang::is_installed("geodata")){
    vnm_lvl3 <- geodata::gadm("VNM", level = 3, path = "vnm_lvl3")
    vnm_lvl3_sf <- sf::st_as_sf(vnm_lvl3)

    vnm_w_centroid <- dart_centroid(vnm_lvl3_sf, 9210)

    # cleanup
    unlink("vnm_lvl3", recursive = TRUE)
}
}
