% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DART_helpers.R
\name{DART_helpers}
\alias{DART_helpers}
\alias{ClearTempDART}
\alias{GetDefaultTopoVars}
\alias{SubsamplePixels}
\alias{does_raster_exist}
\alias{splitDART}
\alias{masterMask}
\alias{raster_mode}
\alias{vec_mode}
\alias{save_diagnostic_plots}
\alias{save_intermediates}
\title{Helper DART functions}
\usage{
ClearTempDART(DART_obj)

GetDefaultTopoVars()

SubsamplePixels(stack0, seed, prop, n_gower)

does_raster_exist(name, dir)

splitDART(p, cellsz = 1000)

masterMask(rasters)

raster_mode(x, na.rm = T)

vec_mode(v, na.rm = T)

save_diagnostic_plots(DART_obj, rast_obj, pdf_fname)

save_intermediates(DART_obj, rast_obj, grd_fname)
}
\description{
Short utility functions specific to DART/webDART functionality
}
\details{
NA
}
\author{
Steve Fick

Travis Nauman

Anna Knight

Brandon McNellis
}
