% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart_max_dist.R
\name{dart_max_dist}
\alias{dart_max_dist}
\title{Calculate maximum distance from centroid of geometry feature to its
boundaries}
\usage{
dart_max_dist(geo_feature, with_centroid = FALSE)
}
\arguments{
\item{geo_feature}{geometry feature (polygon with coordinates)}

\item{with_centroid}{centroid information is included with geo_feature
(default: FALSE)}
}
\value{
A "units"
}
\description{
Calculate maximum distance from centroid of geometry feature to its
boundaries
}
\examples{
\dontrun{
# read in gson vector
hcmc_gson <- st_read(
    readChar(gson_fname, file.info(gson_fname)$size),
    quiet = TRUE
)

# geo features without centroid
get_max_dist(hcmc_gson[1, ])

# geo features comes with centroid
feature_with_centroid <- centroid_with_crs(hcmc_gson[1, ], 9210)
get_max_dist(feature_with_centroid, with_centroid = TRUE)
}
}
