% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_resume_501_tr.R
\name{day_resume_501_tr}
\alias{day_resume_501_tr}
\title{Daily resume of legs}
\usage{
day_resume_501_tr(pattern)
}
\arguments{
\item{pattern}{Character string of the pattern to search in the variable names which ones to use}
}
\value{
A list containing:
\item{darts}{character vector of all the scores of the thrown darts}
\item{1st darts}{character vector of all the scores of the first thrown darts}
\item{2nd darts}{character vector of all the scores of the second thrown darts}
\item{3rd darts}{character vector of all the scores of the third thrown darts}
\item{number of darts}{number of thrown darts}
\item{checkouts}{checkouts score}
\item{180}{number of 180s}
\item{140+}{number of 140+}
\item{100+}{number of 100+}
\item{missed doubles}{number of missed doubles}
\item{busted}{number of busted}
\item{dataset count and \%}{dataset containing counts for each possible dart score}
\item{dataset checkouts}{dataset containing counts for checkouts}
\item{dataset mean and sd}{dataset containing mean and sd values for all 3 darts}
}
\description{
This function is used to create a summary list of all the legs of the day, creating also a dash file
}
\author{
Matteo Miotto
}
