% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{deleteInputFiles}
\alias{deleteInputFiles}
\title{deleteInputFiles}
\usage{
deleteInputFiles(x = "SimulationFiles", ...)
}
\arguments{
\item{x}{\link{SimulationFiles-class} type object.}

\item{...}{Optional arguments `triangles` (bool) to remove triangles files.}
}
\description{
DART input files can be very large. This function deletes those large files
that are not required for post-processing of data in this package.
}
\details{
Delete potentially large input files
}
