% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_persistence.R
\name{dual_line_arrangement}
\alias{dual_line_arrangement}
\title{Line arrangement under point-line duality}
\usage{
dual_line_arrangement(x)
}
\arguments{
\item{x}{output of \code{bigraded_betti}}
}
\value{
Polygons making up the 2-cells of the arrangement.
}
\description{
Computes the cell complex of the line arrangement induced by the multigraded Betti numbers.
}
