% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorial.R
\name{enumerate_lis}
\alias{enumerate_lis}
\title{enumerate_lis}
\usage{
enumerate_lis(x, max_n = Inf)
}
\arguments{
\item{x}{permutation, as an integer vector.}
}
\description{
Enumerates all LIS's of a sequence
}
