% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{execute_schedule}
\alias{execute_schedule}
\title{execute_schedule}
\usage{
execute_schedule(R, V, S, f = NULL)
}
\arguments{
\item{R}{Reduced matrix.}

\item{V}{Upper triangular matrix recording boundaries.}

\item{S}{schedule}
}
\description{
Executes a given schedule of permutations applied to PH-related decomposition.
}
