% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_rips.R
\name{geodesic_dist}
\alias{geodesic_dist}
\title{Geodesic distance}
\usage{
geodesic_dist(x, k = 15L, r = NULL, warn_disconnected = TRUE)
}
\arguments{
\item{x}{point cloud matrix}

\item{k}{number of k-nearest neighbors to form the neighborhood graph with/ See details.}

\item{r}{radius of balls to center at each point to from the neighborhood graph with. See details.}

\item{warn_disconnected}{whether to warn the user if the neighborhood graph is not connected.}
}
\description{
Constructs geodesic distances on a neighborhood graph
}
\details{
TODO
}
