% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{getDate}
\alias{getDate}
\title{Get leg/set/match date of play}
\usage{
getDate(input)
}
\arguments{
\item{input}{A leg1p, leg2p, set1p, set2p, match1p or match2p class object}
}
\value{
Character vector containing the date of play (YYYY-MM-DD)
}
\description{
Get leg/set/match date of play
}
