% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{getID}
\alias{getID}
\title{Get leg/set/match ID}
\usage{
getID(input)
}
\arguments{
\item{input}{A leg1p, leg2p, set1p, set2p, match1p or match2p class object}
}
\value{
Character vector containing leg, set or match ID
}
\description{
Get leg/set/match ID
}
