% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{getPlayers}
\alias{getPlayers}
\title{Get player/s name/s}
\usage{
getPlayers(input)
}
\arguments{
\item{input}{A leg1p, leg2p, set1p, set2p, match1p or match2p class object}
}
\value{
Character vector containing player/s name/s
}
\description{
Get player/s name/s
}
\author{
Matteo Miotto
}
