% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{getStats}
\alias{getStats}
\title{Get leg/set/match stats}
\usage{
getStats(input)
}
\arguments{
\item{input}{A leg1p, set1p or match1p class object}
}
\value{
A list containing leg, set or match stats (means, checkouts, doubles, power scoring, number of darts and number of legs/sets won and lost)
}
\description{
Get leg/set/match stats
}
