% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{getWin}
\alias{getWin}
\title{Win or lost}
\usage{
getWin(input)
}
\arguments{
\item{input}{A leg1p, leg2p, set1p, set2p, match1p or match2p class object}
}
\value{
0 if lost, 1 if won
}
\description{
Win or lost
}
