% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{getWinner}
\alias{getWinner}
\title{Get leg/set/match winner}
\usage{
getWinner(input)
}
\arguments{
\item{input}{A leg2p, set2p or match2p class object}
}
\value{
Character vector containing the name of the winner
}
\description{
Get leg/set/match winner
}
