% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS.R
\name{GIS}
\alias{GIS}
\alias{GetRastData}
\alias{ApplyMasks}
\alias{MakeCandidates}
\alias{MakeTargetData}
\alias{GetTarget}
\title{NA}
\usage{
GetRastData(DART_obj, expanded_target)

ApplyMasks(DART_obj, input_raster_stack)

MakeCandidates(DART_obj, raster_data, target)

MakeTargetData(DART_obj, masked_data, target, i0)

GetTarget(DART_obj, x, y)
}
\description{
NA
}
\details{
NA
}
