% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{greedy_min_cross}
\alias{greedy_min_cross}
\title{Greedy schedule minimization}
\usage{
greedy_min_cross(p, q, lcs, opt = c("minimize", "maximize"), use_lcs = TRUE)
}
\value{
(2 x m) matrix of move permutations.
}
\description{
Creates a schedule of move permutation using a greedy heuristic
}
\details{
Given two permutations, \code{p} and \code{q}
}
