% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{greedy_min_cross2}
\alias{greedy_min_cross2}
\title{Greedy Scheduling}
\usage{
greedy_min_cross2(
  x,
  y,
  lcs = perm_lcs(x, y),
  opt = c("minimize", "maximize"),
  strategy = c("target displacement", "pairwise", "local spearman")
)
}
\description{
Greedy Scheduling
}
