% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_scores.R
\name{high_scores}
\alias{high_scores}
\title{hitting high scores exercise}
\usage{
high_scores(sector)
}
\arguments{
\item{sector}{20 (default), 19 or 18}
}
\description{
This function is used to train hitting high scores sectors
}
\author{
Matteo Miotto
}
