% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagesToDirections.R
\docType{methods}
\name{imagesToDirections,Images,function-method}
\alias{imagesToDirections,Images,function-method}
\title{imagesToDirections}
\usage{
\S4method{imagesToDirections}{Images,`function`}(x, fun = function(x)
  mean(x[x != -1]))
}
\arguments{
\item{\link{Images-class}}{object}
}
\description{
Convert an {Images-class} object to a {Directions-class} object
}
