% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{imagesToDirectionsDF}
\alias{imagesToDirectionsDF}
\title{imagesToDirectionsDF}
\usage{
imagesToDirectionsDF(x, fun)
}
\arguments{
\item{x}{\link{Images-class} object}

\item{fun}{Function to apply across each image.}
}
\value{
data frame
}
\description{
Convert an \link{Images-class} object to a {Directions-class} object
}
\details{
Aggregate images to single values
}
