% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorial.R
\name{kendall_dist}
\alias{kendall_dist}
\title{Kendall distance}
\usage{
kendall_dist(x, y = NULL, normalize = FALSE)
}
\arguments{
\item{x}{a vector of comparable elements.}

\item{y}{a vector of elements of the same kind as \code{x}, or NULL. See details.}

\item{normalize}{whether to normalize the distance to [0,1].}
}
\description{
Measures the kendall distance between two vectors
}
\details{
This functions computes the Kendall distance between arbitrary vectors, which is equivalent 
to counting the number of inversions between the pairs of elements between the two vectors.  
Its expected that both \code{x} and \code{y} have comparable elements, and that 
the type of the elements can be ordered, e.g. \link[base]{sort} and \link[base]{sort} can be applied. 
See below for example inputs. \cr
\cr
If \code{normalize}=\code{TRUE}, the inversion count is normalized to [0,1].\cr
\cr
If \code{y} is not supplied, the Kendall distance from \code{x} to the identity 
permutation of \code{x} is reported.
}
\examples{
x <- sample(seq(50))

## This reports the number of inversions needed to sort x into ascending order
kendall_dist(x)

## The distance can also computed between permutations
kendall_dist(x, sample(x))

## Any type that follows a natural order can be used 
kendall_dist(letters[1:20], sample(letters[1:20]))
}
