% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leg_501.R
\name{leg_501}
\alias{leg_501}
\title{Single leg 501 in a match}
\usage{
leg_501(player1, player2)
}
\arguments{
\item{player1}{player1 name}

\item{player2}{player2 name}
}
\value{
A list containing:
\item{winner}{character vector of the winner}
\item{player1.stat}{Leg stats for player1}
\item{player2.stat}{Leg stats for player2}
}
\description{
This function is used as calculator for a single 501 leg
}
\author{
Matteo Miotto
}
