% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMatch.R
\name{loadMatch}
\alias{loadMatch}
\title{Load Match}
\usage{
loadMatch(file)
}
\arguments{
\item{file}{file path. Could be ommitted and file selection would be promped}
}
\description{
Load a match file
}
\author{
Matteo Miotto
}
