% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTraining501.R
\name{loadTraining}
\alias{loadTraining}
\alias{loadTraining501}
\title{Load 501 training legs}
\usage{
loadTraining(file)
}
\arguments{
\item{file}{file path. Could be ommitted and file selection would be prompted}
}
\description{
Load a 501 training file with different legs in it
}
\author{
Matteo Miotto
}
