% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{move_pair}
\alias{move_pair}
\title{move_pair}
\usage{
move_pair(
  symbol,
  source,
  target,
  lcs,
  rule = c("earliest", "latest", "closest", "furthest", "all")
)
}
\value{
a pair (i,j) of indicies indicating the move.
}
\description{
Constructs valid LCS edit operation
}
\details{
This function constructs a single, particular type of edit operation (a \emph{move}), and is usually called recursively. 
Given a single character \code{symbol} and two strings \code{source} and \code{target} whose longest common subsequence (LCS) is 
is \code{lcs}, this function computes a single edit operation using heuristic \code{rule} such that 
applying the edit operation \code{source} brings it closer to \code{target}, and increases the size of the LCS
by 1.
}
