% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{move_sequence}
\alias{move_sequence}
\title{move_sequence}
\usage{
move_sequence(symbols, s, t, lcs, ordered = FALSE, rule = "all")
}
\description{
Recursively produces sequences of move permutations
}
\details{
Given a set \code{symbols} lying in the set difference \code{s} \ \code{lcs}, this function 
recursively constructs sequences of edit operations which transform \code{s} -> \code{t}. If \code{ordered}
is \code{TRUE}, the edit operations move \emph{symbols} in the order they were given, otherwise all move 
sequences using the given \code{rule} are generated recursively. \cr
\cr 
NOTE: When \code{ordered} is FALSE, the output of this function is exponential in the size of \code{symbols} 
and \code{s}, and thus may be intractable for even modest input sizes.
}
