% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorial.R
\name{perm_lcs}
\alias{perm_lcs}
\title{perm_lcs}
\usage{
perm_lcs(p, q = NULL)
}
\description{
Computes the Longest Common Subsequence (LCS) of a pair of permutations.
}
\details{
If both \code{p} and \code{q} are supplied, the LCS of between them is computed 
and returned. Alternatively if only \code{p} is supplied, the LCS between p and the identity 
permutation is returned. In both cases, computing the LCS reduces to computing the LIS.
}
