% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{pivots}
\alias{pivots}
\title{Pivot entries}
\usage{
pivots(x)
}
\arguments{
\item{x}{a matrix type.}
}
\description{
Returns the pivot entries in a matrix.
}
\details{
The function returns the (row, column) index pairs of the lowest row entries associated 
to each non-zero column of \code{x}, with optimizations used if \code{x} is a sparse matrix.
}
