% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDirections.R
\name{plotDirections}
\alias{plotDirections}
\title{plotDirections}
\usage{
plotDirections(azimuth, zenith, value, azimuthOffsetVal = 0,
  outerRadius = max(zenith) + max(zenith) * 0.01, zenithLabPch = 20,
  zenithLabCol = "darkgrey", zenithLabCex = 1, brks = seq(min(value),
  max(value), length.out = 10), cols = c("dark grey",
  colorRampPalette(c("purple", "blue3", "yellow", "red"))(length(brks) -
  3), "firebrick4"), ...)
}
\arguments{
\item{azimuth}{Numeric. Azimuth angle with DART conventions}

\item{zenith}{Numeric. Zenith angle with DART conventions}

\item{value}{Numeric. Values associated with the given azimuth and zenith angles}

\item{azimuthOffsetVal}{Numeric. Scene offset (degrees) as shown in the DART GUI.}

\item{outerRadius}{Numeric. Maximum radius (degrees) of polar plot}

\item{zenithLabPch}{Numeric. Pch for zenith label.}

\item{zenithLabCol}{Character. Colour for zenith label.}

\item{zenithLabCex}{Numeric. Cex for zenith label.}

\item{brks}{Numeric. Breaks for colour palette e.g. seq(0, 1, by = 0.1). Optional.}

\item{cols}{Character. Colours for given breaks. Optional.}

\item{...}{Additional options passed to points() when drawing directions points.}
}
\description{
Plot directions data as polar plot.
}
\examples{
#Inputs are DART oriented directions (as seen in the DART files and \\link{Directions-class})
plotDirections(azimuth = rep(225, 10),
               zenith = seq(0, 90, length.out = 10),
               value = 1:10)
#Output plot uses 'upward' directions from ground, where e.g.:
  0deg (270deg) azimuth faces north (west)
  0deg (90deg) zenith faces upward (horizon)
}
