% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary.R
\name{print.fbm}
\alias{print.fbm}
\title{Prints filtration boundary matrix objects.}
\usage{
\method{print}{fbm}(x, ...)
}
\arguments{
\item{x}{a filtration boundary matrix object.}

\item{...}{unused.}
}
\description{
Prints filtration boundary matrix objects.
}
