% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_matrix.R
\name{psp_matrix}
\alias{psp_matrix}
\title{psp_matrix}
\usage{
psp_matrix(
  i = NULL,
  j = NULL,
  x = NULL,
  dims,
  type = c("bool", "integer", "numeric")
)
}
\description{
User friendly constructed of a Permutable SParse matrix (psp_matrix). \cr
\cr
A psp_matrix is a sparse matrix representation designed to allow efficient permutation. Like any traditional 
sparse matrix representation (CSC, CSR, etc.), the storage complexity of a psp_matrix is proportional to the 
number of non-zero entries. However, swapping any two rows or columns takes just constant time. With other 
representations, performing a row/column swap in a matrix A can require up to O(nnz(A)) where 
nnz(A) = number of non-zero entries in A.
}
