% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_matrix.R
\name{PspMatrix}
\alias{PspMatrix}
\title{PspMatrix}
\description{
PspMatrix

PspMatrix
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PspMatrix$new()}}
\item \href{#method-print}{\code{PspMatrix$print()}}
\item \href{#method-as.Matrix}{\code{PspMatrix$as.Matrix()}}
\item \href{#method-clone}{\code{PspMatrix$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PspMatrix$new(
  i = NULL,
  j = NULL,
  x = NULL,
  dims,
  type = c("bool", "integer", "numeric")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{row indices of non-zero entries.}

\item{\code{j}}{column indices of non-zero entries.}

\item{\code{x}}{non-zero entries corresponding to positions (\code{i}, \code{j}) in the matrix.}

\item{\code{dims}}{dimension of matrix to construct}

\item{\code{type}}{storage type for the non-zero entries. Deduced from \code{x} if missing. See details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PspMatrix$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as.Matrix"></a>}}
\if{latex}{\out{\hypertarget{method-as.Matrix}{}}}
\subsection{Method \code{as.Matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PspMatrix$as.Matrix(type = "CSC")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{storage type for the non-zero entries. Deduced from \code{x} if missing. See details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PspMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
