% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{r_rips_complex}
\alias{r_rips_complex}
\title{Random Vietoris-Rips complex}
\usage{
r_rips_complex(n, radius, coords = FALSE, ...)
}
\description{
Generates a Rips complex from sampled points on the unit square.
}
