% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinadic.R
\name{rank_lex}
\alias{rank_lex}
\title{rank_lex}
\usage{
rank_lex(x, n)
}
\arguments{
\item{x}{matrix whose columns represent \code{k}-combinations.}

\item{n}{numerator of the binomial coefficient}
}
\value{
integer vector of the positions of the given combinations.
}
\description{
Given a combination \code{x}, computes its position out of all lexicographically-ordered (\code{n} choose \code{2}) combinations.
}
\details{
The mapping is done via an lexicographically-ordered combinadic mapping.
}
\references{
McCaffrey, J. D. "Generating the mth lexicographical element of a mathematical combination." MSDN Library (2004).
}
