% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{rb3DtoNc}
\alias{rb3DtoNc}
\title{rb3DtoNc}
\usage{
rb3DtoNc(x = "SimulationFiles", ...)
}
\arguments{
\item{x}{\link{SimulationFiles-class} type object.}

\item{ncCompressionFactor}{Compression factor (0 - 9) for writing ncdf files (see ncdf4 package)}
}
\value{
\link{SimulationFiles-class} type object.
}
\description{
DART radiative budget .bin files can be very large. This function replaces all
.bin files with .nc files, which can be compressed and are faster to read.
}
\details{
Convert radiative budget .bin to .nc
}
