% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{read_barcodes}
\alias{read_barcodes}
\title{Read bardcodes}
\usage{
read_barcodes(
  x,
  dims = c(0L, 1L),
  values = NULL,
  collapse = TRUE,
  essential = FALSE
)
}
\arguments{
\item{x}{a reduced boundary matrix with column labels, or the output object of \link{reduce}. See details.}

\item{dims}{the homology class dimensions to include in the output.}

\item{values}{optional function values associated with the filtration. If excluded the index persistence is reported. See details.}

\item{collapse}{whether to remove persistent pairs that have 0 persistence. Defaults to true.}

\item{essential}{whether to keep essential classes which do not die. Defaults to false.}
}
\description{
Extracts the persistence diagram from reduced matrices.
}
\details{
Given a reduced boundary matrix \code{x}, this function extracts the persistent pairs
representing points on the persistence diagram. If \code{x} is a matrix, it is expected to have simplices (as strings)
in its \code{colnames} to extract dimension information, otherwise if \code{x} is a list encoding each reduction matrix 
separately the dimension information is inferred from the supplied \code{dims}.
}
