% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{reduce}
\alias{reduce}
\title{Reduction algorithm}
\usage{
reduce(
  D,
  field = c("mod2", "reals"),
  output = c("RV", "RU", "R", "V", "U"),
  indices = NULL,
  options = c(clearing = TRUE),
  validate = TRUE,
  show_progress = FALSE
)
}
\arguments{
\item{D}{either filtration boundary matrix or an object of class "fbm" from e.g. \link[dart]{boundary_matrix}.}

\item{field}{Field of coefficients to use. Defaults to modulo 2 coefficients. See details.}

\item{output}{Type of reduction output requested. See details.}

\item{indices}{Integer indices of which columns to reduce. By default all columns are reduced.}

\item{options}{Options and optimizations which affect how the reduction proceeds.}

\item{show_progress}{whether to print a progress bar as the columns are reduced.}
}
\description{
The standard PH algorithm (and its variants)
}
\details{
This function applies the standard algorithm to decompose a filtration boundary matrix representation
into a variety of outputs.
}
