% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_CRS}
\alias{reference_CRS}
\title{Reference CRS to be used in DART}
\format{
A CRS-class object
}
\usage{
reference_CRS
}
\description{
CRS is from the Colorado_river_watershed_mask ('refrast')
}
\keyword{datasets}
