% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceUse.R
\docType{methods}
\name{resourceUse,SimulationFiles-method}
\alias{resourceUse,SimulationFiles-method}
\title{resourceUse}
\usage{
\S4method{resourceUse}{SimulationFiles}(x = "SimulationFiles")
}
\arguments{
\item{SimulationFiles}{\link{SimulationFiles-class} type object}
}
\description{
Return data frame with simulation name, memory usage and time taken
}
\details{
Get compute resource usage
}
