% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{resourceUse}
\alias{resourceUse}
\title{ResourceUse}
\usage{
resourceUse(x = "SimulationFiles")
}
\arguments{
\item{x}{\link{SimulationFiles-class} type object}
}
\description{
Return a data frame with information on the resource use for a \link{SimulationFiles-class} type object
}
\details{
Return resource use
}
