% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume_all_501_tr.R
\name{resume_all_501_tr}
\alias{resume_all_501_tr}
\title{Overall resumes}
\usage{
resume_all_501_tr()
}
\value{
A list containing:
\item{number of legs}{number of played legs}
\item{darts}{character vector of all the scores of the thrown darts}
\item{1st darts}{character vector of all the scores of the first thrown darts}
\item{2nd darts}{character vector of all the scores of the second thrown darts}
\item{3rd darts}{character vector of all the scores of the third thrown darts}
\item{number of darts}{number of thrown darts}
\item{180}{number of 180s}
\item{140+}{number of 140+}
\item{100+}{number of 100+}
\item{checkouts}{checkouts score}
\item{closing doubles}{doubles that closed the legs}
\item{missed doubles}{number of missed doubles}
\item{missed}{character vector of all missed doubles}
\item{checkout rate}{checkout rate}
\item{busted}{number of busted}
\item{dataset_count_perc}{dataset containing counts for each possible dart score}
\item{dataset_checkouts}{dataset containing counts for checkouts}
\item{dataset_doubles}{dataset containing counts for missed and closing doubles}
\item{dataset_mean_sd}{dataset containing mean and sd values for all 3 darts}
\item{dataset_power}{dataset containing power scoring counts}
}
\description{
This function is used to create a summary list of all the legs played.
}
\author{
Matteo Miotto
}
