% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtration.R
\name{rips_filtration}
\alias{rips_filtration}
\title{Rips filtration}
\usage{
rips_filtration(x, diameter = "default", dim = 1L, ...)
}
\arguments{
\item{x}{either a row-oriented point cloud or a \link[stats]{dist} object.}

\item{diameter}{diameter of balls to place at every point. Defaults to the 2 * < enclosing radius of \code{x} >.}

\item{dim}{maximum dimension to expand the rips complex to do. Defaults to 1, producing the neighborhood graph of \code{x}.}
}
\description{
Rips filtration
}
