% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_clock_3darts.R
\name{round_clock_3darts}
\alias{round_clock_3darts}
\title{round the clock 3 darts}
\usage{
round_clock_3darts(section)
}
\arguments{
\item{section}{which section to score in: "s" for singles (default), "d" for doubles or "t" for trebles}
}
\description{
This function is used to play a modified version of around the clock
}
\author{
Matteo Miotto
}
