% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveMatch.R
\name{saveMatch}
\alias{saveMatch}
\title{Save Match}
\usage{
saveMatch(data, file)
}
\arguments{
\item{data}{match variable}

\item{file}{filename}
}
\description{
Save the match as a file txt or csv
}
\author{
Matteo Miotto
}
