% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{sequenceParameters}
\alias{sequenceParameters}
\title{sequenceParameters}
\usage{
sequenceParameters(x)
}
\arguments{
\item{\link{SimulationFiles-class}}{or \link{SimulationData-class} class object}
}
\description{
return a data frame. A row describes the parameters (parametre*) for a simulation (simName).
}
\details{
Get data frame of all sequence parameters
}
