% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_501.R
\name{set_501}
\alias{set_501}
\title{Single 501 set in a match}
\usage{
set_501(player1, player2, nlegs, which_set, match.id, which_leg_match)
}
\arguments{
\item{player1}{player1 name}

\item{player2}{player2 name}
}
\value{
A list containing:
\item{set.stat}{set statistics}
\item{player1.set.stat}{Set stats for player1}
\item{player2.set.stat}{Set stats for player2}
\item{set.legs.stat}{Legs statistics}
}
\description{
This function is used as calculator for a single 501 set
}
\author{
Matteo Miotto
}
