% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{simulate_moves}
\alias{simulate_moves}
\title{Simulate Moves}
\usage{
simulate_moves(x, schedule, dim)
}
\arguments{
\item{x}{the output of \link{reduce}. Only supports dimension-specific matrices.}

\item{schedule}{Matrix of move indices for the \code{dim} reduced matrices.}

\item{dim}{which homology dimension to apply moves too.}
}
\description{
This function applies a set of moves to a R = DV decomposition.
}
