% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vineyards.R
\name{simulate_vineyard}
\alias{simulate_vineyard}
\title{simulate_vineyard}
\usage{
simulate_vineyard(
  R,
  V,
  S,
  f = NULL,
  D = NULL,
  check_valid = TRUE,
  progress = FALSE
)
}
\arguments{
\item{R}{a square reduced matrix}

\item{V}{a square upper-triangular matrix}

\item{S}{a (2 x m) matrix giving the indices of the transpositions.}

\item{f}{an (optional) function to execute for each decomposition. See details.}

\item{D}{an (optional) boundary matrix \code{R} was derived from, used to verify the decomposition is correct.}
}
\description{
Simulates a vineyard of persistence diagrams
}
