% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_Classes.R
\docType{class}
\name{SimulationData-class}
\alias{SimulationData-class}
\title{Generic SimulationData class}
\description{
Generic SimulationData class that extends to data classes for specific DART products
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame.}
}}

\seealso{
\link{Images-class} \link{Directions-class} \link{RB3D-class}
}
