% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_Classes.R, R/AA_Generics.R
\docType{class}
\name{SimulationFilter-class}
\alias{SimulationFilter-class}
\alias{product<-}
\alias{iters<-}
\alias{bands<-}
\alias{variablesRB3D<-}
\alias{variables<-}
\alias{typeNums<-}
\alias{imageTypes<-}
\alias{imageNos<-}
\title{SimulationFilter class.}
\usage{
product(x) <- value

iters(x) <- value

bands(x) <- value

variablesRB3D(x) <- value

variables(x) <- value

typeNums(x) <- value

imageTypes(x) <- value

imageNos(x) <- value
}
\description{
SimulationFilter class.
}
\section{Slots}{

\describe{
\item{\code{bands}}{character.}

\item{\code{variables}}{character.}

\item{\code{iters}}{character.}

\item{\code{variablesRB3D}}{character.}

\item{\code{typeNums}}{character.}

\item{\code{imageTypes}}{character.}

\item{\code{imageNos}}{numeric.}

\item{\code{product}}{character.}
}}

\seealso{
\code{\link{simulationFilter}}
}
