% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{simulationFilter}
\alias{simulationFilter}
\title{Create \linkS4class{SimulationFilter} class}
\usage{
simulationFilter(product = "character", ...)
}
\arguments{
\item{product}{One of: `directions`, `rb3D`, `images`.}

\item{...}{Optional arguments of: `bands`, `variables`, `iterations`, `variablesRB3D`,
`typeNums`, `imageTypes`, `imageNos`. See \link{SimulationFilter-class} for full
description.}
}
\description{
Function for creating the \linkS4class{SimulationFilter} class
}
\examples{

}
\seealso{
\code{\link{SimulationFilter-class}}
}
