% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_rips.R
\name{sparse_rips}
\alias{sparse_rips}
\title{Sparse rips setup procedure}
\usage{
sparse_rips(x, alpha = "default")
}
\arguments{
\item{x}{matrix of points or 'dist' object.}

\item{alpha}{the maximum scale used to define the neighborhood graph.}
}
\description{
This function compute the information necessary
to efficiently construct sparse rips complexes/filtrations up to scale \code{alpha}
at varying levels of sparsification. Specifically, this function constructs the neighborhood 
graph representing the 1-skeleton of the rips complex up to scale \code{alpha} of \code{x}, 
the greedy permutation of \code{x}, and the covering radii parameters all of which are 
used to define sparse rips complexes and filtrations.
}
