% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_rips.R
\name{sparse_rips_complex}
\alias{sparse_rips_complex}
\title{Sparse rips complex}
\usage{
sparse_rips_complex(x, epsilon = 0.1, alpha = "default", dim = 1L, ...)
}
\arguments{
\item{x}{point cloud matrix, or a 'dist' object.}

\item{epsilon}{degree of approximation. See details.}

\item{alpha}{diameter to build the filtration to. Default to the enclosing radius of the space.}

\item{dim}{highest dimension of simplices to include in the filtration.}
}
\value{
a simplextree object.
}
\description{
Sparse rips complex
}
\references{
http://donsheehy.net/research/cavanna15geometric.pdf
}
