% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_rips.R
\name{sparse_rips_filtration}
\alias{sparse_rips_filtration}
\title{Sparse rips filtration}
\usage{
sparse_rips_filtration(x, epsilon = 0.1, alpha = "default", dim = 1L, ...)
}
\arguments{
\item{x}{point cloud matrix, or a 'dist' object.}

\item{epsilon}{degree of approximation. See details.}

\item{alpha}{diameter to build the filtration to. Default to the enclosing raidus of the space.}

\item{dim}{highest dimension of simplices to include in the filtration.}
}
\value{
a filtration object.
}
\description{
This is a thin wrapper for \code{sparse_rips} and \code{sparsify_rips} which 
are used in tandem to construct a sparse rips filtration. The resulting rips filtration at scale 
\code{alpha} has a persistence diagram which is a (1+\code{epsilon})-approximation of the persistence
diagram of the rips filtration of \code{x} at scale \code{alpha}.
}
\references{
http://donsheehy.net/research/cavanna15geometric.pdf
}
