% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_rips.R
\name{sparsify_rips}
\alias{sparsify_rips}
\title{Rips sparsification procedure}
\usage{
sparsify_rips(
  x,
  p,
  w,
  epsilon,
  alpha,
  dim = 1L,
  prune_vertices = FALSE,
  filtered = FALSE
)
}
\description{
Converts a given rips complex into a sparsified rips complex or filtration
}
\details{
This function is meant to be used on the outputs of \code{sparse_rips} function. 
Given a neighborhood graph \code{x} (as a simplicial complex) with \emph{n} vertices, a greedy permutation 
\code{p} and weight vector \code{w} each of length \emph{n}, this function produces a sparse rips 
complex (or filtration) up to scale \code{alpha} whose persistence diagram is a (1+\code{epsilon}) approximation 
the persistence diagram of \code{x}.
}
