% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorial.R
\name{spearman_dist}
\alias{spearman_dist}
\title{Spearman distance}
\usage{
spearman_dist(x, y = NULL, normalize = FALSE)
}
\arguments{
\item{x}{a vector of comparable elements.}

\item{y}{a vector of elements of the same kind as \code{x}, or NULL. See details.}

\item{normalize}{whether to normalize the distance to [0,1].}
}
\description{
Computes the Spearman's footrule distance.
}
\details{
This function computes the Spearmans footrule or "displacement" distance between arbitrary vectors.
The Spearman distance can be thought of as a type of l_1 distance between vectors, as it effectively just 
measures the sum absolute difference of the ranks of the elements between vectors. 
Its expected that both \code{x} and \code{y} have comparable elements, and that 
the type of the elements can be ordered, e.g. \link[base]{sort} and \link[base]{sort} can be applied. \cr
\cr
If \code{normalize}=\code{TRUE}, the inversion count is normalized to [0,1].\cr
\cr
If \code{y} is not supplied, the Spearman distance from \code{x} to the identity 
permutation of \code{x} is reported.
}
