% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecifyDART.R
\name{SpecifyDART}
\alias{SpecifyDART}
\title{Function to specify a DART object for analysis}
\usage{
SpecifyDART(
  ID,
  mask,
  response,
  PSCS,
  EC,
  other_masks = NULL,
  other_controls = list(),
  target_masks = list(),
  years = c(1984:2018),
  spatial_directory = getwd(),
  results_directory = getwd(),
  intermediate_directory = getwd(),
  use_synthetic_control = F,
  SC_intervention = NULL,
  super_pixel = T,
  fresh = F,
  search_radius = 2000,
  target_radius = numeric(),
  buffer_radius = 90,
  n_matches = 100
)
}
\arguments{
\item{ID}{Identification/name of this DART object. See details.
BEM: maybe make it so ID is a column name in the SPDF???}

\item{mask}{Study area mask. SpatVector.}

\item{response}{List of response variable filenames}

\item{other_masks}{Character vector of other mask filenames to apply}

\item{years}{Years that response variable is desired}

\item{spatial_directory}{Directory path containing spatial files.}

\item{results_directory}{Directory path for results files.}

\item{intermediate_directory}{Directory for intermediate files}

\item{use_synthetic_control}{Should the synthetic control method be used? T/F
Will fail if years < 6 or no intervention year set.}

\item{SC_intervention}{Year of synthetic control intervention.
Required if use_synthetic_control is TRUE}

\item{fresh}{if TRUE, remove all files in results and intermediate directory before run}

\item{search_radius}{Radius in meters to search for reference pixels.}

\item{target_radius}{Radius in meters of target area}

\item{buffer_radius}{Radius in meters of target-reference buffer zone}

\item{n_matches}{Number of topographic matches to search for.
Optional. See details.}

\item{controls}{Character vector specifying files of controls}
}
\description{
This is a wrapper function that initializes a `DART-class` object using parameters
inputs from the user. If an inappropriate parameter is passed, an informative warning is given.
}
\details{
NA
}
