% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart-package.R
\name{str_to_simplex}
\alias{str_to_simplex}
\title{str_to_simplex}
\usage{
str_to_simplex(x, ...)
}
\arguments{
\item{x}{a vector, list, or other iterable of strings to give to sapply.}

\item{...}{passed to sapply.}
}
\description{
str_to_simplex
}
