% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{sunAngles}
\alias{sunAngles}
\title{sunAngles}
\usage{
sunAngles(x = "SimulationFiles")
}
\arguments{
\item{x}{sF \link{SimulationFiles-class}}
}
\value{
data frame
}
\description{
Get sun angles for each simulation
}
