% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_control.R
\name{synthetic_control}
\alias{synthetic_control}
\alias{GetCausalImpact}
\alias{DoParallelImpact}
\alias{DoCausalImpact}
\alias{GS}
\title{Synthetic control functions}
\usage{
GetCausalImpact(DART_obj, var, ID, chunks = 1)

DoParallelImpact(DART_obj, ID, target, control, gower)

DoCausalImpact(DART_obj, ID, target, control, gower)

GS(oi)
}
\description{
These functions run the synethetic control routines for DART-class
objects.

They take place AFTER the ART-pixel selection
}
\details{
NA
}
\author{
Steve Fick

Brandon McNellis
}
