% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{tappToRadiance}
\alias{tappToRadiance}
\title{tappToRadiance}
\usage{
tappToRadiance(x = "SimulationData")
}
\arguments{
\item{x}{\link{SimulationData-class} type object.}
}
\value{
\link{SimulationData-class} type object.
}
\description{
Convert Tapp (K) to Radiance (W m2 sr-1 um-1) using Planck function at the equivalent Band wavelength
}
\details{
Convert Tapp to Radiance
}
