% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DART_Code.R
\name{test.mult}
\alias{test.mult}
\title{Stage 2: Embed multiple testing in the aggregation tree}
\usage{
test.mult(alpha = 0.05, Llist, Dist0, T1)
}
\arguments{
\item{alpha}{(numeric) Vector of desired FDR}

\item{Llist}{(list of matrixes) Aggregation tree presented in the form of matrix list.}

\item{Dist0}{(matrix) Distance matrix}

\item{T1}{(vector) Input Z-values transformed from p-values}
}
\value{
(list of vectors) Index of rejected nodes on each layers
}
\description{
Stage 2: Embed multiple testing in the aggregation tree
}
