% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_501.R
\name{tr_501}
\alias{tr_501}
\title{Single leg 501 training}
\usage{
tr_501()
}
\value{
A list containing:
\item{darts}{character vector of all the scores of the thrown darts}
\item{first_darts}{character vector of all the scores of the thrown first darts}
\item{second_darts}{character vector of all the scores of the thrown second darts}
\item{third_darts}{character vector of all the scores of the thrown third darts}
\item{number of darts}{number of thrown darts}
\item{checkout}{checkout score}
\item{missed doubles}{number of missed doubles}
\item{busted}{number of busted}
}
\description{
This function is used as calculator for a single 501 leg
}
\author{
Matteo Miotto
}
