% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{training}
\alias{training}
\title{Training session}
\usage{
training(save_dir)
}
\arguments{
\item{save_dir}{in which directory are stored the csv files for savings}
}
\description{
This function is used for a single training session
}
\author{
Matteo Miotto
}
