% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training501.R
\name{training501}
\alias{training501}
\title{Single leg 501 training}
\usage{
training501(player, save = T, file = "path/to/file")
}
\description{
This function is used as calculator for a single 501 leg in training
}
\author{
Matteo Miotto
}
