% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{versionInfo}
\alias{versionInfo}
\title{ResourceUse}
\usage{
versionInfo(x)
}
\arguments{
\item{x}{\link{SimulationFiles-class} type object}
}
\description{
Return a data frame with information on the resource use for a \link{SimulationFiles-class} type object
}
\details{
Return resource use
}
