% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webDART.R
\name{webDART}
\alias{webDART}
\alias{RunWebDART}
\title{WebDART interface}
\usage{
RunWebDART(directory = getwd(), debugMessage = FALSE)
}
\description{
NA
}
\details{
NA
}
\author{
Steve Fick

Anna Knight

Samuel Chambers

Brandon McNellis
}
