% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{accessors}
\alias{accessors}
\alias{simname}
\alias{files}
\alias{bands}
\alias{iters}
\alias{variables}
\alias{variablesRB3D}
\alias{typeNums}
\alias{imageTypes}
\alias{imageNums}
\title{Access object information}
\usage{
product(x)

simname(x)

files(x)

bands(x)

iters(x)

variables(x)

variablesRB3D(x)

typeNums(x)

imageTypes(x)

imageNums(x)
}
\arguments{
\item{x}{\linkS4class{SimulationFilter} or \linkS4class{SimulationFiles} class}
}
\description{
Generic functions to access information from the objects with classes defined in this package
}
\examples{
sF <- simulationFilter(product = "directions")
bands(sF)

\dontrun{
#access information within SimulationFiles object
#define the simulation directory
simDir <- "C:/Users/<Username>/DART/user_data/simulations/cesbio/"
simFiles <- getFiles(simDir)
#show bands that are selected
bands(simFiles)
#show 'type numbers' that have been selected
typeNums(simFiles)
}

}
