% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary.R
\name{boundary_matrix}
\alias{boundary_matrix}
\title{Boundary matrix}
\usage{
boundary_matrix(
  x,
  dim = "all",
  labeled = TRUE,
  field = c("mod2", "integer", "real")
)
}
\arguments{
\item{x}{list of simplices, simplex tree, or a filtration.}

\item{dim}{dimension of boundary matrix desired}

\item{labeled}{whether to add simplex labels to the corresponding matrices.}

\item{field}{intended}
}
\description{
Constructs filtration boundary matrices
}
\details{
This function computes and returns the elementary chains associated with a filtered simplicial complex.
The filtration is expected to be given as list of a simplices.
}
