% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart-package.R
\docType{package}
\name{dart-package}
\alias{dart}
\alias{dart-package}
\title{dart: Companion R package for the DART (Dengue Advanced Readiness Tools) project by OUCRU}
\description{
This package provides functions that performs raster spatial analysis and processing to be used in the DART (Dengue Advanced Readiness Tools) project by OUCRU, in collaboration with RMIT University Vietnam. The main function of this package calculates the average of raster pixel values in a buffer around a geometric feature (i.e. Average rainfall of a commune/ward in a city).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tuyenhn/dart}
  \item Report bugs at \url{https://github.com/tuyenhn/dart/issues}
}

}
\keyword{internal}
