% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart-package.R
\docType{package}
\name{dart}
\alias{dart}
\alias{dart-package}
\title{dart: a package containing tools for simulating persistent homology dynamically.}
\description{
This packages provides a set of tools for computing and updating the matrix 
	decomposition which characterizes the persistent homology of a filtered complex.
}
\keyword{internal}
