% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart_centroid.R
\name{dart_centroid}
\alias{dart_centroid}
\title{sf::st_centroid() wrapper with optional CRS transformation}
\usage{
dart_centroid(sf, target_crs = 4326)
}
\arguments{
\item{sf}{sf object}

\item{target_crs}{CRS which the centroid calculation will be performed
with (optional)}
}
\value{
original sf object with new centroid column
}
\description{
sf::st_centroid() wrapper with optional CRS transformation
}
\examples{
\dontrun{
# read in gson vector
hcmc_gson <- st_read(
    readChar(gson_fname, file.info(gson_fname)$size),
    quiet = TRUE
)

hcmc_gson_w_centroid <- dart_centroid(hcmc_gson, 9210)
}
}
