% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart_max_dist.R
\name{dart_max_dist}
\alias{dart_max_dist}
\title{Calculate maximum distance from centroid of geometry feature to its
boundaries}
\usage{
dart_max_dist(geo_feature, centroid)
}
\arguments{
\item{geo_feature}{geometry feature (polygon with coordinates)}

\item{centroid}{centroid information column}
}
\value{
A "units"
}
\description{
Calculate maximum distance from centroid of geometry feature to its
boundaries
}
\examples{
\dontrun{
# read in gson vector
hcmc_gson <- st_read(
    readChar(gson_fname, file.info(gson_fname)$size),
    quiet = TRUE
)

feature_with_centroid <- centroid_with_crs(hcmc_gson[1, ], 9210)
get_max_dist(feature_with_centroid, feature_with_centroid$centroid)
}
}
