% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{deleteFiles}
\alias{deleteFiles}
\title{deleteFiles}
\usage{
deleteFiles(x = "SimulationFiles", ...)
}
\arguments{
\item{x}{\link{SimulationFiles-class} type object.}

\item{...}{\code{trianglesInput} remove "triangles" input files? (bool)}

\item{...}{\code{maketOutput} remove "maket.txt" output file? (bool)}
}
\description{
DART input files can be very large. This function deletes those large files
that are not required for post-processing of data in this package.
}
\details{
Delete potentially large input files
}
