% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_persistence.R
\name{dual_graph}
\alias{dual_graph}
\title{Dual graph}
\usage{
dual_graph(arrangement)
}
\arguments{
\item{arrangement}{geometry collection of polygons (output from \code{dual_line_arrangement}).}
}
\value{
edge list of the dual graph.
}
\description{
Computes the dual graph of a line arrangement.
}
