% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{getData}
\alias{getData}
\title{Main function: get DART data}
\usage{
getData(x, sF)
}
\arguments{
\item{x}{simulation directory or directories (character) or \link{SimulationFiles-class} object}

\item{sF}{\link{SimulationFilter-class} if x = \code{character}}
}
\description{
Main function to get data from DART simulation outputs in a friendly 'long' data format that is part of
an object that extends a \link{SimulationData-class} type object
}
