% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{getFiles}
\alias{getFiles}
\title{Get DART output filenames}
\usage{
getFiles(x = "character", sF = "SimulationFilter")
}
\arguments{
\item{x}{simulation directory or directories (character)}

\item{sF}{\link{SimulationFilter-class} object}

\item{...}{Optional arguments of: \code{nCores}: number of cores to use when loading data.}
}
\description{
Function for getting \link{SimulationFiles-class} type object. Useful to perform a 'dry run' of
\link{getData} by exploring the files that will vary based on the contents of \code{x} and
the configuration of \code{sF}.
}
