% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_501.R
\name{match_501}
\alias{match_501}
\title{Single 501 match}
\usage{
match_501(player1, player2, sets2win = 1, nlegs = 1)
}
\arguments{
\item{player1}{player1 name}

\item{player2}{player2 name}
}
\value{
A list containing:
\item{match.stat}{match statistics}
\item{sets.stat}{sets statistics}
\item{player1.set.stat}{Set stats for player1}
\item{player2.set.stat}{Set stats for player2}
\item{legs.stat}{Legs statistics}
}
\description{
This function is used as calculator for a single 501 match
}
\author{
Matteo Miotto
}
