% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{removeRelief}
\alias{removeRelief}
\title{removeRelief}
\usage{
removeRelief(x = "RB3D", DEM = "RasterLayer")
}
\arguments{
\item{x}{\link{RB3D-class} type object.}

\item{DSM}{\code{RasterLayer} type object with height above ground level (m) and - preferably - a finer
horizontal reoslution than that of the radiative budget cells in x. The center of the DSM must be georeferenced
to the center of the radiarive budget data in x. The DSM can have a larger extent than x.}
}
\description{
Remove underlying orography from a \link{RB3D-class} dataset using a digital elevation
model (DEM) of class \code{RasterLayer} that is georeferenced to \link{RB3D-class}.
}
\details{
Remove underlying orography
}
