% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{sequenceParameters}
\alias{sequenceParameters}
\title{sequenceParameters}
\usage{
sequenceParameters(x)
}
\arguments{
\item{\link{SimulationFiles-class}}{or \link{SimulationData-class} class object}
}
\value{
data.frame()
}
\description{
Return a data frame where rows describe a parameter (parametre*) for a simulation (simName).
}
\details{
Get data frame of all sequence parameters
}
