% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{simulationFilter}
\alias{simulationFilter}
\title{Create \linkS4class{SimulationFilter} class}
\usage{
simulationFilter(product = "character", ...)
}
\arguments{
\item{product}{One of "directions", "rb3D", "images".}

\item{...}{}
}
\description{
Function for creating the \linkS4class{SimulationFilter} class
}
\examples{

}
\seealso{
\code{\link{SimulationFilter-class}}
}
