% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{simulationFilter}
\alias{simulationFilter}
\title{Create \linkS4class{SimulationFilter} class}
\usage{
simulationFilter(product = "character", x, ...)
}
\arguments{
\item{product}{One of: `directions`, `rb3D`, `images`.}

\item{x}{\link{SimulationFiles-class} object if product is missing.}
}
\value{
\linkS4class{SimulationFilter} type object
}
\description{
Function for creating the \linkS4class{SimulationFilter} class. Define a product, then
Optional arguments of: `bands`, `variables`, `iterations`, `variablesRB3D`,
`typeNums`, `imageTypes`, `imageNos`. See \link{SimulationFilter-class} for full
description.
}
\examples{
sF <- daRt::simulationFilter(product = "images",
                             bands = as.integer(0:2),
                             iters = c("ITER1", "ITER2"),
                             variables = "BRF",
                             imageNos = as.integer(c(5, 7)),
                             imageTypes = c("ima", "ima_transmittance"))

}
\seealso{
\code{\link{SimulationFilter-class}}
}
