% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{versionInfo}
\alias{versionInfo}
\title{versionInfo}
\usage{
versionInfo(x)
}
\arguments{
\item{x}{\link{SimulationFiles-class} object}
}
\description{
Get the version used for the given simulation data
}
\details{
Simulation version info
}
