% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vineyards.R
\name{vineyards_schedule}
\alias{vineyards_schedule}
\title{vineyards_schedule}
\usage{
vineyards_schedule(K0, K1, w0 = NULL, w1 = NULL)
}
\description{
Creates a schedule of transpositions needs to transform one filtration into another.
}
\details{
This function assumes K0 and K1 contain the same elements.
}
