% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Generics.R
\name{wavelengths}
\alias{wavelengths}
\title{wavelengths}
\usage{
wavelengths(x = "SimulationFiles")
}
\arguments{
\item{x}{sF \link{SimulationFiles-class}}
}
\value{
data frame
}
\description{
Get full information on wavelengths for each band
}
